// src/types/BucklerCharFrame.ts
type BucklerCharFrame = {
  frame: {
    x: number;
    y: number;
    w: number;
    h: number;
  };
  rotated: boolean;
  trimmed: boolean;
  spriteSourceSize: {
    x: number;
    y: number;
    w: number;
    h: number;
  };
  sourceSize: {
    w: number;
    h: number;
  };
  anchor: {
    x: number;
    y: number;
  };
}

export default BucklerCharFrame;